Create user context to handle auth state
